requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.11
click==8.1.8
coloredlogs==15.0.1
cryptography==45.0.3
deprecated==1.2.18
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.115.9
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.82.1
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pdfminer-six==20250327
pdfplumber==0.11.6
pillow==11.2.1
posthog==4.2.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pypdfium2==4.30.1
pypika==0.48.9
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
starlette==0.45.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.22.0
